Improve error reporting in command.py
Previously, if an error occurred when constructing the Popen object,
Command.run would throw its own exception because self.process would not
be defined. This commit makes a small modification so that the traceback
from the error is instead returned from Command.run, which will no
longer throw an exception in this situation.

+++ app/helpers/command.py
@@ -47,6 +47,7 @@ class Command(object):
         timeout = False
         if thread.is_alive():
             timeout = True
-            self.process.terminate()
-            thread.join()
+            if self.process:
+                self.process.terminate()
+                thread.join()
         return timeout, self.output, self.error
